switching to integer storage of money

+++ app/models/expense.rb
@@ -5,5 +5,5 @@ class Expense < ApplicationRecord
 
   validates :expense_tracker, :amount, :date, presence: true
   validates :amount, numericality: { greater_than_or_equal_to: 0,
-                                     less_than_or_equal_to: BigDecimal(10**8) }
+                                     less_than_or_equal_to: Integer(10**8) }
 end
