+++ app/models/statement.rb
@@ -5,7 +5,7 @@ class Statement < ApplicationRecord
 
   validates :account, :balance, :date, presence: true
   validates :balance, numericality: { greater_than_or_equal_to: 0,
-                                      less_than_or_equal_to: BigDecimal(10**8) }
+                                      less_than_or_equal_to: Integer(10**8) }
   validates :notes, allow_nil: true, format: { with: /\A[\w\s[[:punct:]]]+\z/ }
 
   scope :sorted_by_date, -> { order(date: :desc) }
