+++ spec/models/statement_spec.rb
@@ -14,7 +14,7 @@ RSpec.describe Statement, type: :model do
     it {
       should validate_numericality_of(:balance)
         .is_greater_than_or_equal_to(0)
-        .is_less_than_or_equal_to(BigDecimal(10**8))
+        .is_less_than_or_equal_to(Integer(10**8))
     }
 
     # balance
